requirements.txt
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.1.3
python-dotenv==1.0.1
requests==2.31.0
